style(model): format HF reader and model loader code
- Apply consistent formatting per .clang-format
- Break long return statements for readability
- Remove unnecessary blank line in model_loader.cpp
- Remove 'Phase 1 收敛版' suffix from model_loader_design.md title

+++ include/aethermind/model/formats/hf/hf_safetensors_index.h
@@ -21,7 +21,7 @@ struct HfSafetensorEntry {
     RawTensorView view{};
 
     AM_NODISCARD size_t ByteSize() const noexcept {
-        return static_cast<size_t>(data_offset_end - data_offset_begin);
+        return data_offset_end - data_offset_begin;
     }
 };
